requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
confluent-kafka==2.3.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
httpx==0.25.2
pyyaml==6.0.1

